requirements.txt
fastapi
uvicorn[standard]
openai
chromadb
sentence-transformers
pdfplumber
PyMuPDF
python-docx
openpyxl
langchain-text-splitters
python-dotenv
sqlalchemy
python-multipart
